Sample Code for C#
Codes for the following exercises:
Console Application Assignment
In Visual Studio, create a console application that does the following:

Takes an input from the user, multiplies it by 50, then prints the result to the console. (Note: make sure your code can take inputs larger than 10,000,000).

Takes an input from the user, adds 25 to it, then prints the result to the console.

Takes an input from the user, divides it by 12.5, then prints the result to the console.

Takes an input from the user, checks if it is greater than 50, then prints the true/false result to the console.

Takes an input from the user, divides it by 7, then prints the remainder to the console (tip: think % operator).

Add comments to each line or block of your code to explain what it does exactly, so that another developer could read and understand your code.